+++ packages/console-types/src/types/catalog/item-manifest.ts
