+++ packages/console-types/src/types/catalog/item.ts
